Move impl MethodBodyTypeChecker to method_body_typechecker module
