+++ compiler-lib/src/type_checking/mod.rs
@@ -20,7 +20,6 @@
 
 pub mod builtin_types;
 pub mod checker;
-pub mod expr_typechecker;
 pub mod method_body_typechecker;
 pub mod type_analysis;
 pub mod type_system;
